refactor: enum for error messages
